worked on debugging wikipedia

+++ graph/src/types.rs
@@ -9,11 +9,11 @@ use std::str::FromStr;
 /// Type used to index the Nodes.
 pub type NodeT = u32;
 /// Type used to index the Node Types.
-pub type NodeTypeT = u16;
+pub type NodeTypeT = u32;
 /// Type used to index the Edges.
 pub type EdgeT = u64;
 /// Type used to index the Edge Types.
-pub type EdgeTypeT = u16;
+pub type EdgeTypeT = u32;
 /// Type used for the weights of the edges.
 pub type WeightT = f32;
 /// Type used for the parameters of the walk such as the return weight (p),
